UPD: regression module migration
