+++ mapie/regression/__init__.py
@@ -1,0 +1,9 @@
+from .regression import MapieRegressor
+from .quantile_regression import MapieQuantileRegressor
+from .time_series_regression import MapieTimeSeriesRegressor
+
+__all__ = [
+    "MapieRegressor",
+    "MapieQuantileRegressor",
+    "MapieTimeSeriesRegressor"
+]
