+++ mapie/tests/test_time_series_regression.py
@@ -13,12 +13,15 @@ from typing_extensions import TypedDict
 from mapie._typing import NDArray
 from mapie.aggregation_functions import aggregate_all
 from mapie.metrics import regression_coverage_score
+from mapie.regression import MapieTimeSeriesRegressor
 from mapie.subsample import BlockBootstrap
-from mapie.time_series_regression import MapieTimeSeriesRegressor
 
+random_state = 1
 X_toy = np.array(range(5)).reshape(-1, 1)
 y_toy = (5.0 + 2.0 * X_toy ** 1.1).flatten()
-X, y = make_regression(n_samples=500, n_features=10, noise=1.0, random_state=1)
+X, y = make_regression(
+    n_samples=500, n_features=10, noise=1.0, random_state=random_state
+)
 k = np.ones(shape=(5, X.shape[1]))
 METHODS = ["enbpi"]
 UPDATE_DATA = ([6], 17.5)
@@ -36,7 +39,9 @@ STRATEGIES = {
     "jackknife_enbpi_mean_ab_wopt": Params(
         method="enbpi",
         agg_function="mean",
-        cv=BlockBootstrap(n_resamplings=30, n_blocks=5, random_state=1),
+        cv=BlockBootstrap(
+            n_resamplings=30, n_blocks=5, random_state=random_state
+        ),
     ),
     "jackknife_enbpi_median_ab_wopt": Params(
         method="enbpi",
@@ -44,13 +49,15 @@ STRATEGIES = {
         cv=BlockBootstrap(
             n_resamplings=30,
             n_blocks=5,
-            random_state=1,
+            random_state=random_state,
         ),
     ),
     "jackknife_enbpi_mean_ab": Params(
         method="enbpi",
         agg_function="mean",
-        cv=BlockBootstrap(n_resamplings=30, n_blocks=5, random_state=1),
+        cv=BlockBootstrap(
+            n_resamplings=30, n_blocks=5, random_state=random_state
+        ),
     ),
     "jackknife_enbpi_median_ab": Params(
         method="enbpi",
@@ -58,7 +65,7 @@ STRATEGIES = {
         cv=BlockBootstrap(
             n_resamplings=30,
             n_blocks=5,
-            random_state=1,
+            random_state=random_state,
         ),
     ),
 }
@@ -252,10 +259,10 @@ def test_linear_regression_results(strategy: str) -> None:
 def test_results_prefit() -> None:
     """Test prefit results on a standard train/validation/test split."""
     X_train_val, X_test, y_train_val, y_test = train_test_split(
-        X, y, test_size=1 / 10, random_state=1
+        X, y, test_size=1 / 10, random_state=random_state
     )
     X_train, X_val, y_train, y_val = train_test_split(
-        X_train_val, y_train_val, test_size=1 / 9, random_state=1
+        X_train_val, y_train_val, test_size=1 / 9, random_state=random_state
     )
     estimator = LinearRegression().fit(X_train, y_train)
     mapie_ts_reg = MapieTimeSeriesRegressor(estimator=estimator, cv="prefit")
@@ -271,7 +278,10 @@ def test_results_prefit() -> None:
 
 def test_not_enough_resamplings() -> None:
     """Test that a warning is raised if at least one residual is nan."""
-    with pytest.warns(UserWarning, match=r"WARNING: at least one point of*"):
+    with pytest.warns(
+        UserWarning,
+        match=r"WARNING: at least one point of*"
+    ):
         mapie_ts_reg = MapieTimeSeriesRegressor(
             cv=BlockBootstrap(n_resamplings=1, n_blocks=1), agg_function="mean"
         )
@@ -353,3 +363,33 @@ def test_MapieTimeSeriesRegressor_beta_optimize_eeror() -> None:
         mapie_ts_reg._beta_optimize(
             alpha=0.1, upper_bounds=X, lower_bounds=X_toy
         )
+
+
+def test_deprecated_path_warning() -> None:
+    """
+    Test that a warning is raised if import with deprecated path.
+    """
+    with pytest.warns(
+        FutureWarning,
+        match=r".*WARNING: Deprecated path*"
+    ):
+        from mapie.time_series_regression import MapieTimeSeriesRegressor
+        _ = MapieTimeSeriesRegressor()
+
+
+def test_consistent_class() -> None:
+    """
+    Test that importing a class with a new or obsolete path
+    produces the same results.
+    """
+    from mapie.regression import MapieTimeSeriesRegressor as C2
+    from mapie.time_series_regression import MapieTimeSeriesRegressor as C1
+
+    mapie_c1 = C1(random_state=random_state).fit(X, y)
+    mapie_c2 = C2(random_state=random_state).fit(X, y)
+
+    y_pred_1, y_pis_1 = mapie_c1.predict(X, alpha=0.1)
+    y_pred_2, y_pis_2 = mapie_c2.predict(X, alpha=0.1)
+    np.testing.assert_allclose(y_pis_1[:, 0, 0], y_pis_2[:, 0, 0])
+    np.testing.assert_allclose(y_pis_1[:, 1, 0], y_pis_2[:, 1, 0])
+    np.testing.assert_allclose(y_pred_1, y_pred_2)
